VAR models

The VAR model we use in this case is of the following form: $x_t = ν + A_1 (x_traj_{t-1}', p_traj_{t-1}', h(x_traj_{t-1}, p_traj_{t-1})')' + ... + A_p (x_traj_{t-p}', p_traj_{t-p}', h(x_traj_{t-p}, p_traj_{t-p})')'$

where $x_traj_t$ is the trajectory of the observables at time $t$, $p_traj_t$ is the trajectory of the parameters at time $t$ and $h$ is a function that allows for non-linearities in the model.

All relevant information about the VAR model is stored in the struct:

Destabilisation.VARmodel_module.VARmodel — Type

VARmodel

An element of the struct VARmodel contains all the information needed to define a VAR model.

Fields

h::Union{Function,Nothing}: function that computes the regressor of the form (x_t',p_t',h(x_t,p_t)')'

p::Int64: order of the VAR model

d_x::Int64: dimension of the observable variable

d_p::Int64: dimension of the parameter variable

d_h::Int64: dimension of the hidden variable

d_y::Int64: dimension of the augmented variable

B_hat::Matrix{Float64}: estimated coefficient matrix

Σ_hat_u::Matrix{Float64}: estimated covariance matrix of the residuals (unbiased estimator)

Σ_tilde_u::Matrix{Float64}: estimated covariance matrix of the residuals (maximum likelihood)

Σ_β_hat::Matrix{Float64}: estimated covariance matrix of the coefficient matrix (unbiased estimator)

Γ_hat::Matrix{Float64}: estimator for YY'/T

Σ_x1_hat::Matrix{Float64}: estimated MSE matrix of the one step forecast (unbiased estimator)

The following function can be used to predict the next observable values given a VAR model and the current values of the observables and parameters:

Destabilisation.VARmodel_module.oneStepPred — Function

oneStepPred(model::VARmodel,x_traj::Matrix;p_traj::Union{Matrix,Nothing}=nothing)

Compute the one step prediction of the VAR model given by model for the trajectory x_traj and the parameter trajectory p_traj. If p_traj is not given, it is assumed that the VAR model is time invariant.