Fitting a VAR model to observed data

The VAR model we use in this case is of the following form: $x_t = ν + A_1 (x_traj_{t-1}', p_traj_{t-1}', h(x_traj_{t-1}, p_traj_{t-1})')' + ... + A_p (x_traj_{t-p}', p_traj_{t-p}', h(x_traj_{t-p}, p_traj_{t-p})')'$

where $x_traj_t$ is the trajectory of the observables at time $t$, $p_traj_t$ is the trajectory of the parameters at time $t$ and $h$ is a function that allows for non-linearities in the model.

A suitable order $p$ for the VAR model can be determined using the function:

Destabilisation.VARorder — Function

VARorder(traj::Matrix; p_max::Int64=10, criterion::String="AIC")

This function computes the optimal order of a VAR model for a given trajectory x_traj by minimizing the AIC, FPE, HQ or SC criterion. The optimal order is returned as an integer. The p_max argument denotes the maximal order which is considered. If the system is forced by varying parameter values the respective time series should be given in the p_traj argument which must have the same number of datapoints as the x_traj argument. The h argument allows for non-linearities to be included in the model. To this end h must be a vector valued function which takes two arguments x and p if p_traj !== nothing and one argument x if p_traj === nothing. So in total the optimal VAR order of a model of the form x_traj_t = ν + A_1 (x_traj_{t-1}', p_traj_{t-1}', h(x_traj_{t-1}, p_traj_{t-1})')' + ... + A_p (x_traj_{t-p}', p_traj_{t-p}', h(x_traj_{t-p}, p_traj_{t-p})')' is computed.

After selecting a suitable order the VAR model can be fitted to the data using the function:

Destabilisation.fitVARmodel — Function

fitVARmodel(x_traj::Matrix; p_traj::Union{Matrix,Nothing}=nothing, h::Union(Function,Nothing)=nothing, p::Int64=1)

Fits a VAR(p) model of the form x_traj_t = ν + A_1 (x_traj_{t-1}', p_traj_{t-1}', h(x_traj_{t-1}, p_traj_{t-1})')' + ... + A_p (x_traj_{t-p}', p_traj_{t-p}', h(x_traj_{t-p}, p_traj_{t-p})')' to the given input data. It returns a VARmodel object which contains the estimated parameters and the estimated covariance matrices.