Format Test

These functions can be used to verify if a given object has indeed the desired format needed for the respecive application.

Destabilisation.FormatTests.check_traj — Function

check_traj(traj)

Checks if traj has the format of a trajectory i.e. if traj is a two dimentional matrix with all entries given by numbers. Further it is always assumed that each column of trajectory represents a datapoint of the time series, i.e. each row denotes the time evolution of a single variable.

Destabilisation.FormatTests.check_DynamicalSystem — Function

check_DynamicalSystem(progressor,observable,v_init,p_init,p_bounds,v_length,p_length,x_length)

Compatibility test for initialisation of an object of DynamicalSystem.

check_DynamicalSystem(progressor,observable,v_init,p_init,p_bounds)

Compatibility test for initialisation of an object of DynamicalSystem.

Destabilisation.FormatTests.check_h — Function

check_h(h::Function, x_traj::Union{Matrix{Float64},Nothing}, p_traj::Union{Matrix{Float64},Nothing})

Compatibility test for the function h used for the inclusion of non-linearities in the model. h is supposed to be given by a Vector{Float64} valued function which takes one or two vector arguments, depending on whether a timeseries of parameters is included or not.

Destabilisation.FormatTests.check_xph — Function

check_xph(x_traj::Matrix,p_traj::Union{Matrix,Nothing}=nothing,h::Union{Function,Nothing}=nothing)

Compatibility test for simultaneously checking the trajectory x_traj, the parameter trajectory p_traj and the function h. x_traj, p_traj and h are supposed to satify their individual compatibility tests and in addition, the number of datapoints in x_traj and p_traj must be the same. Besides checking the compatibility of x_traj, p_traj and h, the function also returns the dimensions of the trajectory x_traj, the parameter trajectory p_traj and the output of h at the first datapoint as d_x, d_p and d_h, respectively.