Integrate trajectories

Destabilisation.integrateTraj — Function

integrateTraj(DS::DynamicalSystem, g::Function, T::Int64; include_reference_traj::Bool=false)

This function iteratively computes the evolution of the state variables v, the parameters p and the observables.

Arguments

DS::DynamicalSystem: Element of the type DynamicalSystem which encodes the dynamics of the system under considerantion

g::Function: Function determining the one-step evolution of the parameter values. I.e. g must be such that it takes one Argument p::Vector{Float64} containing the current parameter values and returns the new parameter values again in the form of Vector{Float64}

T::Int64: Number of time steps

include_reference_traj::Bool=false: If true, the reference trajectory is included in the output

Returns

`varr': Vector of dimension (DS.lengthv,T) containing the time evolution of the state space variable

`parr': Vector of dimension (DS.lengthp,T) containing the time evolution of the parameters

`xarr': Vector of dimension (DS.lengthx,T) containing the time evolution of the observables

v_ref_arr': This is only returned ifincludereferencetraj==true. Vector of dimension (DS.length_v,T) containing the v values which are obtained by computing the next timestep using the past v value as given invarrbut the initial parameter valuesDS.pinit. IfDS.randomveclength !== nothing` then the same noise is used for both integrations.

x_ref_arr': This is only returned ifincludereferencetraj==trueand is obtained by applying theobservablefunction of theDynamicalSystemobject tovrefarrandDS.p_init`.

noise': This is only returned ifDS.randomveclength !== nothing`. Vector of dimension (DS.randomveclength,T) containing the noise values which are used for the integration of the reference trajectory.

integrateTraj(DS::DynamicalSystem, p_traj::Matrix{Float64}; include_reference_traj::Bool=false)

This function iteratively computes the evolution of the state variables v and the observables x of a system forced by a given trajectory of the parameters p. The trajectories for v, p and x are always returned as the first there return values. If include_reference_traj is set to true, the function also returns additional trajectories of v and x for fixed parameters which in each timestep are computed from the v values of the trajectories with varying parameters, allowing for a comparison of the trajectories with varying and fixed parameters. If the field random_vec_length of the DynamicalSystem is not nothing, then the same noise is used for both the trajectory with variing parameters and the one with fixed parameters and the function additionaly returns the noise trajectory. This is also the case if the field random_vec_length is not nothing while include_reference_traj is set to false.

integrateTraj(DS::DynamicalSystem, T::Int64)

Integrates the trajectory of a system for T time steps with constant parameters and returns the trajectory of the state variables v, the parameters p and the observables. If the field random_vec_length of the DynamicalSystem is not nothing, the function also returns the noise vector containing the noise trajectory used to integrate the system.