Matrices

Here some usefull functions are presented which return matrices which are for instance needed to fit a model to the recored data.

Destabilisation.Y — Function

Y(traj::Matrix; p::Int64=1, T::Int64=100)

This function turns a trajectory traj which is a matrix with K rows into a matrix Y as defined in Lütkepohls book in equation (3.2.1). If T+p excedes the length of the trajectory, i.e. traj has less than T+p columns, then Y is retruned with T = size(traj)[2]-p. If on the other hand the trajectory is longer than T+p then Y given by slicing the last T columns out of traj.