Test Parameter Dependence

The following function can be used to find the locations of the parameter dependent coefficients in the regressor of a given VAR model

Destabilisation.getParamLocations — Function

get_param_locations(model::VARmodel, param_idx::Int64)

Returns the indices of the observables that depend on the parameter with index param_idx in the form of an array which can be used as the argument param_locations in the function C.

In order to determine if the observables significantly depend on the parameters, the following function can be used

Destabilisation.testParamCausality — Function

testParamCausality(model::VARmodel, param_idx::Int64, T::Int64; siglvl::Float64 = 0.01)

Returns a tuple (is_causal, p_value) where is_causal is a boolean indicating whether or not the parameter with index param_idx is causal for the observables in the VAR model model and p_value is the p-value of the test. In order to compute the test statistic it is necessary to give the length T of the timeseries, that was used to estimate the VAR model. The test is performed at significance level siglvl.