Test estimated VAR models

Here we pesent some tests that we can perform on estimated VAR models in order to check their validity.

In this first test we present a function that quantifies the validity of a model by comparing the one step predictions of the model with the actual values of the observables by performing a hypothesis test.

Destabilisation.testPredictions — Function

testPredictions(model::VARmodel,
    x_traj::Matrix;
    p_traj::Union{Matrix,Nothing}=nothing,
    siglvl::Float64=0.05)

This function tests the validity of the given VAR model by checking the prediction error of the model. The prediction error is computed as the sum of the squared nomalized, uncorrlated prediction errors of the model. The prediction error is then compared to the critical value of the χ² distribution with degrees of freedom equal to the number of observations times the number of variables. If the percentil of the prediction error is smaller than the significance level siglvl, the model is rejected. The function returns a tuple consisting of a boolean value whether the model is accepted or not and the percentil of the prediction error.

The whiteness i.e. the absence of autocorrelation in the residuals of the model is tested by the following function.

Destabilisation.LMtest — Function

LMtest(model::VARmodel,h::Int64,x_traj::Matrix;p_traj::Union{Matrix,Nothing}=nothing,siglvl::Float64=0.05)

This function is the implementation of the Lagrange multiplier test for testing the whiteness of the residuals. Here x_traj and p_traj are the trajectories used to estimate the model. h determines up to how many lags we test for uncorrelated residualts. siglvl is the significance level of the test.