Determine time scales of VAR models

The following method is used to determine the time scales of a VAR model. The time scales are determined by the eigenvalues of the VAR model.

Destabilisation.timeScale — Function

timeScale(model::VARmodel)

Determines a proxy of the slowest timescale of the VAR model model by returning the absolute value of the smallest root of the characteristic polynomial of the VAR model. This is done by fitting a polynomial of degree deg to the function f(z) = det(Id - sum([z^i * B_i for i in 1:p])) where B_i is the i-th block of the matrix B_hat of the VAR model model. The root of this polynomial with the smallest absolute value (which should be larger than one for the model to be stable) is returned. Note that this function can only be applied to VAR models without parameter variables and without hidden variables.

timeScale(B_hat::Matrix{Float64})

Auxiliary function that returns the timescale for a given matix B_hat

timeScale(progressor::Function, 
            v_init::Vector{Float64}, 
            p_init::Vector{Float64}; 
            T_init::Int64 = 200,
            T_max::Int64 = 10000,
            rel_error::Float64 = 1e-1)

Determines a proxy of the slowest timescale of the dynamical system defined by the progressor function progressor with given initial state variables v_init and parameter variables p_init. This is done by integrating the dynamical system for a time T_init and then doubling the time until the error on the time scale approximation is smaller than rel_error * abs(ts - 1). The function then returns the time scale approximation and the error on the approximation.

timeScale(progressor::Function, 
            v_init::Vector{Float64}, 
            p_init::Vector{Float64}; 
            T_init::Int64 = 200,
            T_max::Int64 = 10000,
            rel_error::Float64 = 1e-1)

Determines a proxy of the slowest timescale of the observable dynamics of the the dynamical systen DS. This is done by integrating the dynamical system for a time T_init and then doubling the time until the error on the time scale approximation is smaller than rel_error * abs(ts - 1). The function then returns the time scale approximation and the error on the approximation.

This function uses the following auxiliary function which turns a given function into a polynomial

Destabilisation.toPolynomial — Function

toPolynomial(f::Function; deg::Int64 = 1, tolerance::Float64 = 1e-4)

Fit a polynomial of degree deg to the function f. The polynomial is fitted using the least squares method. The accuracy of the polynomial is checked by comparing the function f to the polynomial on 100 equidistant points in an appropriately chosen interval. If the maximum error is larger than tolerance, a warning is issued. Finally the polynomial is returned.