Matrices

Here some usefull functions are presented which return matrices which are for instance needed to fit a model to the recored data.

Destabilisation.slice_traj — Function

slice_traj(data::Matrix; p::Int64=1, T::Int64=100)

Slices out the last T+p columns from the data argument. Here p denotes the number of presamples and T encodes the number of actual samples.

Destabilisation.Y — Function

Y(traj::Matrix; p::Int64=1, T::Union{Int64,Nothing}=nothing)

This function turns a trajectory traj, which is a matrix with K rows into a matrix, Y as defined in Lütkepohls book in equation (3.2.1) by slicing the last T colomns from the traj argument. If T===nothing then T is set to the number of columns of traj minus p i.e. the traj argument is interpreted in such a way that it consists of all the presamples plus all the actual samples. If on the other hand T is given and T+p is not equal to the length of the trajectory, i.e. traj has less than T+p columns, an error is throuwn.

As always `p` denotes the number of presamples and `T` encodes the number of actual samples.

Destabilisation.Z_t — Function

Z_t(traj::Matrix, t; p::Int64=1)

Returns the vector Zₜ as defined e.g. in equation (3.2.1) in Lütkepohls book. Here it is assumed that the trajectory traj includes the presample datapoints i.e. yₜ from the book corresponds to the column of traj with index t+p.

As always `p` denotes the number of presamples.

Destabilisation.Z — Function

Z(traj; p::Int64=1, T::Union{Int64,Nothing}=nothing)

Returns Z as defined in Lütkepohls book in equation (3.2.1). If the number of columns of the trajectory traj does not equal T+p then an error is throuwn.

As always p denotes the number of presamples and T encodes the number of actual samples.

Destabilisation.Id — Function

Id(d)

Returns a d by d sparse identity matrix

Destabilisation.J — Function

J(K::Int64, p::Int64)

Returns the matrix J as defined in equation (2.1.11)