Matrices

Here some usefull functions are presented which return matrices which are for instance needed to fit a model to the recored data.

Destabilisation.slice_traj — Function

slice_traj(data::Matrix; p::Int64=1, T::Int64=100)

Slices out the last T+p columns from the data argument. Here p denotes the number of presamples and T encodes the number of actual samples.

Destabilisation.X — Function

X(traj::Matrix; p::Int64=1, T::Union{Int64,Nothing}=nothing)

This function turns a trajectory traj, which is a matrix with K rows, into a matrix X as defined in Lütkepohls book in equation (3.2.1) (where it is called Y) by slicing the last T colomns from the traj argument. If T===nothing then T is set to the number of columns of traj minus p i.e. the traj argument is interpreted in such a way that it consists of all the presamples plus all the actual samples. If on the other hand T is given and T+p is not equal to the length of the trajectory, i.e. traj has less than T+p columns, an error is throuwn. As always p denotes the number of presamples and T encodes the number of actual samples.

Destabilisation.Y_t — Function

Y_t(traj::Matrix, t; p::Int64=1)

Returns the vector Yₜ as defined e.g. in equation (3.2.1) in Lütkepohls book (where it is called Zₜ). Here it is assumed that the trajectory traj includes the presample datapoints i.e. yₜ from the book corresponds to the column of traj with index t+p. As always p denotes the number of presamples.

Destabilisation.Y — Function

Y(traj; p::Int64=1, T::Union{Int64,Nothing}=nothing)

Returns Y as defined in Lütkepohls book in equation (3.2.1) (where it is called Z). If the number of columns of the trajectory traj does not equal T+p then an error is throuwn. As always p denotes the number of presamples and T encodes the number of actual samples.

Destabilisation.Id — Function

Id(d)

Returns a d by d sparse identity matrix

Destabilisation.J — Function

J(K::Int64, p::Int64)

Returns the matrix J as defined in equation (2.1.11)

Destabilisation.create_y_traj — Function

create_y_traj(x_traj::Matrix{Float64};p_traj::Union{Matrix{Float64},Nothing}=nothing,h::Union{Function,Nothing}=nothing)

This function creates a trajectory y_traj from x_traj, p_traj and h s.t. the i-th column of y_traj is given by (x_traj[:,i]', p_traj[:,i]', h(x_traj[:,i],p_traj[:,i])')'.

Destabilisation.F_i — Function

F_i(i::Int64, T::Int64)

Returns the matrix Fᵢ as defined in equation (4.4.2)

Destabilisation.F — Function

F(h::Int64, T::Int64)

Returns the matrix Fᵢ as defined in equation (4.4.2)

Destabilisation.C — Function

C(param_locations::Vector{Int64}, d_x::Int64, d_y::Int64, p::Int64)

C is the matrix as defined in section 3.6.1 in Lütkepohls book. It can be used to test if the observables in the fitted VAR model show a dependence on a given parameter p. To this end the positions of the p dependent entries of y = (x',p',h(x,p)')' must be known and encoded in the param_locations argument. The argument d_x denotes the dimension of the x vector, d_y the dimension of the y vector and p the number of presamples. The function then returns the respective matrix C of dimentions length(param_locations) × d_x * (p * d_y + 1) as a sparse matrix.