Parameter Series

This method is used to generate a series parameter values.

Destabilisation.parameterSeriesGenerator — Function

parameterSeriesGenerator(ts::Float64, p_init::Vector{Float64}, noise::Vector{Float64}; n::Union{Int64,Nothing} = nothing)

Generate a parameter series given the time scale of the system, the initial parameter values as well as the noise amplitudes. The series has the form of a VAR(1) process with the given time scale and the given noise amplitudes varying around the given initial parameter values. The length of the parameter series is determined by the time scale if not specified.

Arguments

ts::Float64: The time scale of the system which for instance can be determined using the timeScale function.

p_init::Vector{Float64}: The initial parameter values.

noise::Vector{Float64}: The noise amplitudes.

n::Union{Int64,Nothing} = nothing: The length of the parameter series. If not specified the length is determined by the time scale.